requirements.txt
flask~=2.3.2
flask-restx
Flask-SQLAlchemy
Flask-Migrate
mysql-connector
mysql-connector-python
pymysql
cryptography
pyjwt~=2.7.0
boto3
SQLAlchemy~=2.0.16
Werkzeug~=2.3.6
alembic~=1.11.1